UserView

+++ src/main/java/com/EAD/EAD_monolithic/dto/DeliveryDTO.java
@@ -1,7 +1,5 @@
 package com.EAD.EAD_monolithic.dto;
 
-import com.EAD.EAD_monolithic.entity.Order;
-import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @Data
 public class DeliveryDTO {
-    private long deliveryId;
+    private int deliveryId;
     private int orderId;
 //    private User customerId;
 //    private User deliveryPersonID;
